third_party/blink/web_tests/platform/generic/http/tests/xmlhttprequest/post-arraybuffer-data-view-expected.txt

Test for ArrayBuffer POST in XMLHttpRequest send

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

xhr.readyState = 4: responseURL = http://127.0.0.1:8000/xmlhttprequest/resources/post-echo-as-ascii.cgi
PASS status is "200"
PASS responseText is "0 1 2 25 45 58 255"
PASS successfullyParsed is true

TEST COMPLETE
